Unify two Rdelta computations
